requirements.txt
pandas==1.4.0
python-dotenv==0.19.2
tweepy==4.5.0
